style: add notification fontSize
