+++ client/src/components/index.js
@@ -1,0 +1,1 @@
+export { default as FormInputContainer } from './FormInputContainer';
